Fix #88 Don't skip "\ESC...\STX" sequences
For ANSI-capable terminals, this reverses commit f827f10145380ec9e685a72e4b8a7a27ef6717ea of 11 September 2010 (about 5 years before the November 2015 Update to Windows 10 made the native consoles on Windows 10 ANSI-capable).

It uses the same logic as the ansi-terminal package to determine if the terminal is ANSI-capable.